解决iOS已下载词典情况下无法离线使用问题
将开启时对词典更新状态的检查从阻塞(await)改为非阻塞(promise)写法

+++ miniprogram/pages/words/words.js
@@ -164,22 +164,23 @@ Page({
     wx.hideLoading()
 
     // 预备“朗读”功能
-    try {
-      this.InnerAudioContext = wx.createInnerAudioContext()
-      this.InnerAudioContext.src = 'https://dict.youdao.com/dictvoice?audio=' + dictionary[index]._id
-      this.InnerAudioContext.onEnded(() => {
-        app.globalData.timeout = setTimeout(this.onPlay, 1000)
-      })
-    } catch(e) {
-      console.log(175, e)
-      this.setData({
-        noAudio: true
-      })
-    }
     if (app.globalData.offline) {
       this.setData({
         noAudio: true
       })
+    } else {
+      try {
+        this.InnerAudioContext = wx.createInnerAudioContext()
+        this.InnerAudioContext.src = 'https://dict.youdao.com/dictvoice?audio=' + dictionary[index]._id
+        this.InnerAudioContext.onEnded(() => {
+          app.globalData.timeout = setTimeout(this.onPlay, 1000)
+        })
+      } catch(e) {
+        console.log(175, e)
+        this.setData({
+          noAudio: true
+        })
+      }
     }
   },
 
@@ -228,18 +229,20 @@ Page({
     }
     
     // 更新“朗读”内容
-    this.InnerAudioContext.destroy()
-    this.InnerAudioContext = wx.createInnerAudioContext()
-    let fontRes = await this.calFontSize(this.data.dictionary[this.data.index].deris)
-    console.log("fontRes: ", fontRes)
-    this.setData({
-      showPlay: true,
-      fontRes: fontRes
-    })
-    this.InnerAudioContext.src = 'https://dict.youdao.com/dictvoice?audio=' + this.data.dictionary[this.data.index]._id
-    this.InnerAudioContext.onEnded(() => {
-      app.globalData.timeout = setTimeout(this.onPlay, 1000)
-    })
+    if (!this.data.noAudio) {
+      this.InnerAudioContext.destroy()
+      this.InnerAudioContext = wx.createInnerAudioContext()
+      let fontRes = await this.calFontSize(this.data.dictionary[this.data.index].deris)
+      console.log("fontRes: ", fontRes)
+      this.setData({
+        showPlay: true,
+        fontRes: fontRes
+      })
+      this.InnerAudioContext.src = 'https://dict.youdao.com/dictvoice?audio=' + this.data.dictionary[this.data.index]._id
+      this.InnerAudioContext.onEnded(() => {
+        app.globalData.timeout = setTimeout(this.onPlay, 1000)
+      })
+    }
   },
 
   // “朗读”与“暂停”
